MB-18433 - OnEvent() callback option added
The OnEvent() option is intended to allow advanced applications to be
able to implement complex functionality under application control,
such as the ability to pause batch ingest and/or merging, in order to
provide backpressure to allow slow persistence to catch up.

Also, updated the tests to utilize the OnEvent() callback, allowing
for non-sleep-based tests and allowing for "go test -race" to pass.

Change-Id: Ifdd3dc0ace27015b5e399234d2914020953a8490
Reviewed-on: http://review.couchbase.org/61128
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>