found-by-smat - acquire lock during Close() steps
smat tests found a potential race situation where all goroutines were
asleep.  The cause is theorized to be a race between the app goroutine
trying to collection.Close() and the persister goroutine.  The
proposed solution is the closer should acquire the collection lock
before closing the stopCh and awaking the various condition variables.

I believe this can be counted as smat's first trophy from moss bugs.

Change-Id: Ib5b9193cd4b884b7d370bd9d407d5861987f71c2
Reviewed-on: http://review.couchbase.org/64516
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>