faster merge through rougher estimation of memory needed
Previously, the merger would use an iterator to estimate the bytes
needed, which was relatively accurate but slower.

This change just sums up all the bytes used by all the to-be-merged
segments, which can wildly over-estimate in the case when the
application is mutating just a single key, but is correct for the
worst-case when the application is always inserting brand new keys.

Change-Id: Ifcab0bd114860ef1a44969e17f9775523a121b9f
Reviewed-on: http://review.couchbase.org/64663
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>