MB-20397 - refresh mmap's on older footers/snapshots
This change introduces a ref-counting wrapper (mmapRef) for mmap
handles.

When a mmapRef's ref-count drops to 0, its memory-mapped slice will be
unmapped and its corresponding file handle (FileRef) will be
DecRef()'ed.

This change also introduces ref-counting to footers.  Footers no
longer directly reference a FileRef, but now reference instead a
mmapRef.  (And, an mmapRef, in turn, references a FileRef.)

Finally, to fix MB-20397, whenever a new footer is appended to a store
file, any older, previous, in-use footers are refreshed to point to
the latest mmapRef.  The hope is that slow readers that are holding
onto a snapshot (and hence, an older footer) can be all migrated to
use the latest mmap'ed handles.  Older mmap handles can then be
dropped sooner rather than later, hopefully relieving pressure on the
VMM.

Change-Id: I69263b4a94bbb26a677f87400066eed039d47ce2
Reviewed-on: http://review.couchbase.org/66476
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>