cleanup middle of footer chain, not just the end
The previous code used to cleanup only the tail end of the footer
chain.  So, if there were some unused footers in the middle of the
chain somewhere, they would be inefficiently left in place.

This change splices out unused footers no matter where they are in the
footer chain.

Change-Id: I38ed810cce84f16faddfe20a5f55494d6bbf4948
Reviewed-on: http://review.couchbase.org/66544
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>