MB-20494 - compacted file is fsync'ed before file-footer written
This change is with regards to fsync and compaction, and previously,
an fsync would occur only _after_ the file-footer in the compacted
file was written.

This change invokes fsync in addition _before_ the file-footer in the
compacted file is written.

Also, the StorePersistOption.NoSync option is propagated to the
compaction codepaths so that it can be consulted as to whether to
fsync should be invoked during compaction.

Change-Id: I21b4224ec97f03642c6a437a293372e15eef120e
Reviewed-on: http://review.couchbase.org/67465
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>