optimize Iterator.SeekTo() by first trying some naive Next()'s
This optimization is based on the assumption that a forward SeekTo()
will often be to a key that's nearby.

In that case, it's fruitful to try some limited number of Next()'s in
a loop, and then devolve back to binary search if the seekToKey isn't
encountered.

In one application benchmark (bleve-query on a wiki dataset with a
query-string of "+text:afternoon +text:coffee"), the search throughput
jumped from 858 queries/sec to 1222 queries/sec after this change.

Change-Id: Id15506426a4ff3b9c40f43f91a54e269bdb7a254
Reviewed-on: http://review.couchbase.org/67930
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>