MB-21059 - fix iteratorSingle to hold a ref to its mmap'ed resource.
Previously, iterators would be associated with their mmap'ed resource,
which helped ensure the mmap'ed resource would remain ref-counted
while the iterator was still in-use.

When an iterator was closed, too, the iterator would also close its
associated mmap'ed resource (decrementing ref-count by 1).

But, when the iteratorSingle optimization was introduced, this
association incorrectly didn't happen with iteratorSingle's, because
the previous code incorrectly performed a type equality check against
just the *iterator struct type (incorrectly forgetting to check the
iteratorSingle struct type).

This fix introduces the InitCloser interface, which is implemented by
both the regular iterator type and the optimized iteratorSingle type.

Change-Id: I71a7ada4439d4777d887634bfbf510d5945290a3
Reviewed-on: http://review.couchbase.org/68255
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>