MB-21477 - consider compaction only when there's >1 segment
A bug introduced in commit f7cc198cb4 around optimizing/fixing
compaction behavior turned out to overly, aggressively compact,
especially in the case when there was only a single segment.  Even
when mossStore reopened a file, for example, which was already
compacted down to a single segment, the previous code would
incorrectly decide that it was time again to compact.

Change-Id: I0beb37122fea8488b4daea6bfee763841d3faf21
Reviewed-on: http://review.couchbase.org/69218
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>