MB-21695 - mossStore reuses existing mmap()'ed regions
Before this change, when mossStore appended a new footer to the end of
the file, it would re-mmap() the entire file.  As more footers are
appended, the mmap()'ed region would grow larger and larger...

    [1111]
    [1111--2222]
    [1111--2222--3333]

Eventually, a huge, contiguous region of virtual address space would
be needed, which the OS might not be able to easily provide.

Instead, with this change, previous mmap()'ed regions are reused by
bumping ref-counts on the older regions, and only the latest appended
section of the file is mmap()'ed.

    [1111]
          [2222]
                [3333]

So, the OS only needs to find contiguous virtual address space for
smaller, more granular regions.

This change also simplifies the code by removing the previous
complexity of updating older snapshots with the latest "full-file"
mmap(), where the code that previously maintained a linked-list of
older footers is now removed.

Change-Id: Id12666e6c99086ba59d6dd6934e92824a061b6bf
Reviewed-on: http://review.couchbase.org/70368
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>