MB-23187 - race detector fix by updating stats after sorting
Spawn the goroutine that updates histograms only after a segment has
been sorted... e.g., after the segment is considered immutable, so
concurrent reads are safe.

As part of this, newBatch() signature is changed to take a pointer to
the root collection (instead of a collection name param which wasn't
being used).

Additionally, a segment that comes from a batch (as opposed to loaded
from storage) holds onto a reference to its root collection so that
the segment can later can invoke rootCollection.updateStats().

Change-Id: I5dccf9509a79f99722d7897cedba37a1af4722f5
Reviewed-on: http://review.couchbase.org/74753
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>