FindStartKeyInclusivePos returns as soon as exact match is found
Previously the comparison only checked for < or >= and would
always continue trying to tighten the lower bound, even if
we'd already found an exact match (part of >= handling).  This
change now explicitly checks for exact matches first, and in
this case, returns immediately with the index that was being
considered.

Change-Id: I8351331c197dc11f289ada7f1cca44767d69a198
Reviewed-on: http://review.couchbase.org/74794
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>