introduce new FindKeyPos method to Segment interface
FindKeyPos is responsible for doing exact key position lookups,
whereas FindStartKeyInclusivePos does a >= lookup.

Previosly, if you wanted to do a Get operation on a particular
key, first FindStartKeyInclusivePos was called, then the
position returned would be passed to GetOperationKeyVal.  But,
if the key didn't exist, we potentially did extra work to
retrieve a key/value pair that isn't the one we wanted.

In the current segment implementation this extra cost wasn't
significant.  But, in more complex segment implementations,
there may be some non-negligible cost.

Change-Id: I11e258ed8b72dbf23363ebaa6fff86b060193514
Reviewed-on: http://review.couchbase.org/74795
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>