MB-24702: Delete old compacted file in separate go routine
The last one to close the old compacted file deletes it.
Now if the old file is huuuge, then it could stall the caller
for a non-trivial amount of time.
The caller could be the query reader or persister.
This stall is seen in DGM index building graphs.
To avoid blocking, have the remove run in a separate go routine.

Change-Id: I6dcdece499adf45ef8a88ac7261536c5df8f168f
Reviewed-on: http://review.couchbase.org/79148
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>