MB-24680: When idle try compactions to speed queries
If the collection merger is idle that means mutations have
either slowed or stopped. At this point it would be better
to trigger an "idle run" which attempts to persist with an
empty higher Snapshot. This will trigger incremental/full
compaction, thereby reducing the number of persisted segments
to speed up queries.

Change-Id: I71f8fa1b8135b800c2800d702e9920b922382a2a
Reviewed-on: http://review.couchbase.org/80163
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>