MB-25419: Don't try to wake merger if it isn't asleep
There was a fix made at (http://review.couchbase.org/60986)
to have the persister wake up the merger if mutations have
stopped but merger is asleep. However, this wake up logic
did not check to see if the merger is indeed asleep.
As a result, when persistence is really fast, we can have
a number of pings sent out to an active merger, causing it
to wake up to just process empty messages/creating garbage.
Simple fix can avoid spurious wakes and is an optimization.

Change-Id: I88b32f1892fd27205056056143962ccef2162a08
Reviewed-on: http://review.couchbase.org/81708
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>