MB-29664 - provide 'base' seg-stack during partial compactions
If the application is configured to use partial compaction and was
also using 'merge operators', then there could be data loss as the
merge-operators were incorrectly not provided values from older or
lower 'base' segments.

IMPORTANT / NOTE -- Even with this commit, there is a remaining issue
with child collections and merge-operators, in the face of a partial
compactions.  Until there is a fix for that, for now applications who
are using merge-operators in their child-collections should configure
moss to only perform full compactions instead of partial compactions.

Change-Id: I81060b01ba77e32c3c8b126eb45fc4b6b0b446af
Reviewed-on: http://review.couchbase.org/94262
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sundar Sridharan <sundar@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>